+++ sdk/containerregistry/azure-containerregistry/samples/sample_delete_tags_async.py
@@ -7,19 +7,18 @@
 # --------------------------------------------------------------------------
 
 """
-FILE: sample_set_image_properties_async.py
+FILE: sample_delete_tags_async.py
 
 DESCRIPTION:
-    This sample demonstrates setting an image's properties on the tag so it can't be overwritten during a lengthy
-    deployment.
+    This sample demonstrates deleting all but the most recent three tags for each repository.
 
 USAGE:
-    python sample_set_image_properties_async.py
+    python sample_delete_tags_async.py
 
     Set the environment variables with your own values before running the sample:
     1) CONTAINERREGISTRY_ENDPOINT - The URL of you Container Registry account
 
-    This sample assumes your registry has a repository "library/hello-world" with image tagged "v1",
+    This sample assumes your registry has at least one repository with more than three tags,
     run load_registry() if you don't have.
     Set the environment variables with your own values before running load_registry():
     1) CONTAINERREGISTRY_ENDPOINT - The URL of you Container Registry account
@@ -32,11 +31,12 @@ USAGE:
 import asyncio
 import os
 from dotenv import find_dotenv, load_dotenv
+from azure.containerregistry import ArtifactTagOrder
 from azure.containerregistry.aio import ContainerRegistryClient
-from samples.utilities import load_registry, get_authority, get_audience, get_credential
+from utilities import load_registry, get_authority, get_audience, get_credential
 
 
-class SetImagePropertiesAsync(object):
+class DeleteTagsAsync(object):
     def __init__(self):
         load_dotenv(find_dotenv())
         self.endpoint = os.environ.get("CONTAINERREGISTRY_ENDPOINT")
@@ -44,27 +44,28 @@ class SetImagePropertiesAsync(object):
         self.audience = get_audience(self.authority)
         self.credential = get_credential(self.authority, is_async=True)
 
-    async def set_image_properties(self):
+    async def delete_tags(self):
         load_registry()
-        # Instantiate an instance of ContainerRegistryClient
         async with ContainerRegistryClient(self.endpoint, self.credential, audience=self.audience) as client:
-            # Set permissions on the v1 image's "latest" tag
-            await client.update_manifest_properties(
-                "library/hello-world",
-                "v1",
-                can_write=False,
-                can_delete=False
-            )
-            # After this update, if someone were to push an update to `<registry endpoint>\library\hello-world:v1`,
-            # it would fail. It's worth noting that if this image also had another tag, such as `latest`,
-            # and that tag did not have permissions set to prevent reads or deletes, the image could still be
-            # overwritten. For example, if someone were to push an update to `<registry endpoint>\hello-world:latest`
-            # (which references the same image), it would succeed.
+            # [START list_repository_names]
+            async for repository in client.list_repository_names():
+                print(repository)
+                # [END list_repository_names]
+
+                # Keep the three most recent tags, delete everything else
+                tag_count = 0
+                async for tag in client.list_tag_properties(
+                    repository, order_by=ArtifactTagOrder.LAST_UPDATED_ON_DESCENDING
+                ):
+                    tag_count += 1
+                    if tag_count > 3:                        
+                        print(f"Deleting {repository}:{tag.name}")
+                        await client.delete_tag(repository, tag.name)
 
 
 async def main():
-    sample = SetImagePropertiesAsync()
-    await sample.set_image_properties()
+    sample = DeleteTagsAsync()
+    await sample.delete_tags()
 
 
 if __name__ == "__main__":
