+++ sdk/containerregistry/azure-containerregistry/samples/sample_hello_world_async.py
